Implemented the functionality for additional points joker
It will still require some work to do. Now it does not disable the joker buttons after answering, so all of them can be used while the answer is selected. Moreover, it adds points available for this question, not the points gained by the player (after time algorithm implemented by Som), but I think it's a cool feature to have, it makes this joker stronger.